Update README.md
# PRODIGY_TrackCode_1

This project was developed as a part of the **Prodigy InfoTech Internship Program**. The goal of this internship is to help students and aspiring developers gain real-world experience by working on hands-on projects under expert guidance.

## About the Project

This repository contains the implementation for the task titled **"TrackCode 1"**, which is part of the internship task track. The objective of this task was to apply fundamental concepts in web development/software development and demonstrate practical skills through a working project.

## Internship Credit

This project is officially submitted for evaluation as part of the [Prodigy InfoTech](https://prodigyinfotech.dev/) Internship Program. I would like to extend my gratitude to the Prodigy InfoTech team for providing this learning opportunity and helping young developers grow their technical skills.